+++ src/tiny-slider-helper.ie8.js
@@ -1,0 +1,12 @@
+/**
+  * tiny-slider
+  * @version 0.3.1
+  * @author William Lin
+  * @license The MIT License (MIT)
+  * @github https://github.com/ganlanyuan/tiny-slider/
+  * 
+  */
+
+// @codekit-prepend "../bower_components/fix-ie/src/es5-arrays.js";
+// @codekit-prepend "../bower_components/fix-ie/src/IE8-child-elements.js";
+// @codekit-prepend "../bower_components/go-native/src/components/IE8-addEventListener.js";