+++ src/tiny-slider.js
@@ -1,0 +1,25 @@
+/**
+  * tiny-slider
+  * @version 0.3.1
+  * @author William Lin
+  * @license The MIT License (MIT)
+  * @github https://github.com/ganlanyuan/tiny-slider/
+  * 
+  */
+
+// @codekit-prepend "../bower_components/fix-ie/src/token-list.js";
+// @codekit-prepend "../bower_components/fix-ie/src/es5-methods.js";
+// @codekit-prepend "../bower_components/requestAnimationFrame/requestAnimationFrame.js";
+// @codekit-prepend "../bower_components/Units/Length.js";
+// @codekit-prepend "../bower_components/go-native/src/components/es5-object.keys.js";
+// @codekit-prepend "../bower_components/go-native/src/components/gn.js";
+// @codekit-prepend "../bower_components/go-native/src/components/optimizedResize.js";
+// @codekit-prepend "../bower_components/go-native/src/components/extend.js";
+// @codekit-prepend "../bower_components/go-native/src/components/isInViewport.js";
+// @codekit-prepend "../bower_components/go-native/src/components/getSupportedProp.js";
+// @codekit-prepend "../bower_components/go-native/src/components/indexOf.js";
+// @codekit-prepend "../bower_components/go-native/src/components/isNodeList.js";
+// @codekit-prepend "../bower_components/go-native/src/components/append.js";
+// @codekit-prepend "../bower_components/go-native/src/components/wrap.js";
+
+// @codekit-prepend "tiny-slider.native.js";