+++ changelog.md
@@ -1,8 +1,10 @@
 #Changelog
 
 #### v1.0.0 beta
+- Totally rewrite with a simple function, co-work with [go-native](https://github.com/ganlanyuan/go-native).
 - Removed: `offset`.
 - Fixed: an issue that navs are still active when slides cann't fill their parent.
+- Fixed: an issue with `items`, when container width is less than one slide in fixedWidth slider.
 
 #### v0.3.1
 - Improved: performance.