+++ changelog.md
@@ -2,8 +2,11 @@
 
 #### v1.0.0 beta
 - Totally rewrite with a simple function, co-work with [go-native](https://github.com/ganlanyuan/go-native).
+- Better [accessibility](https://www.w3.org/WAI/EO/Drafts/tutorials/sliders/) support for people using keyboard navigation or using screen readers. Thanks [DouglasdeMoura](https://github.com/DouglasdeMoura) for the idea. [[issue 4]](https://github.com/ganlanyuan/tiny-slider/issues/4)
 - Improved: optimize scroll and resize events.
 - Added: `autoHeight` option.
+- Renamed options: `nav` => `controls`, `navText` => `controlsText`, `navContainer` => `controlsContainer`, `dots` => `nav`, `dotsContainer` => `navContainer`.
+- Renamed class names: `tiny-current` => `current`, `tiny-visible` => `visible`, `tiny-active` => `active`, `tiny-hide` => `tiny-hidden`, class `disabled` => attribute `disable`.
 - Removed: `offset`.
 - Fixed: an issue that navs are still active when slides cann't fill their parent.
 - Fixed: an issue with `items`, when container width is shorter than one slide in fixedWidth slider.